remove sql defaults for consistent output
in my local doom my sql results have a different color and colons on the side.  hh liked the look of this.  Comparing the header args, my local config does not wrap the results in an example block and does not give the results as code. 

I removed the code from the main result blocks and added verbatim to the sql ones to try to get a more consistent look in our blocks